evince: pass -Dmultimedia=disabled if !supportMultimedia
The evince derivation has an option supportMultimedia?true, which if
set to false will exclude gstreamer from the nativeBuildInputs.
Unfortunately this is not enough; the build will fail with:

  Run-time dependency gstreamer-base-1.0 found: NO (tried pkgconfig)

  meson.build:236:0: ERROR: Dependency "gstreamer-base-1.0" not found, tried pkgconfig

Let's pass "-Dmultimedia=disabled" to meson in this case so that the
build succeeds.

Checked with nixpkgs-hammer, which did not report any new warnings
introduced by this commit.