home-assistant: fix packageOverrides ordering
As documented by lib.composeManyExtensions, overrides intended to be applied later should come last in the list. This commit switches the order to make sure that custom overrides are in fact evaluated after the default ones. Also, since the previous version was basically a reimplementation of composeManyExtensions anyway, switching to that makes it a bit cleaner.