sapling: remove intermediate package
The build for the sapling package happens in three steps:

1. build sapling-isl
2. build sapling-main, copying files from sapling-isl
3. build sapling, copying files from sapling-main

I don't see a reason for splitting sapling-main and sapling. The problem
with splitting is that it makes it hard to use a package override to
patch sapling-main.

Merge sapling-main and sapling so I can easily apply patches with my
nixpkgs/config.nix.