netdata-go.d.plugin: from 0.32.3 to 0.50.0
Netdata package has been updated several times in the meantime, but this
(somewhat crucial) dependency was left forgotten and unloved. So this is
quite huge jump in versions, but it is needed as many of the collectors
were added there (the one I have noticed that is missing is PostgreSQL
collector, but I bet that there is more).

It also adds missing Bash `buildInput` as I noticed that it is needed by
some shell collectors (cgroups I believe).