auto-multiple-choice: change default module dir (fix #214724)
AMC will look for its libraries on the nix store path if and only iff
its default modules directory doesn't exist.  The value for this
variable was /lib/, which exists on some systems.  This commit changes
it to /nonexistent, the "canonical non-existent home directory" per
Debian policy, which probably won't exist anywhere.