tsm-client: 8.1.22.0 -> 8.1.23.0
"Update History" (release notes):
https://www.ibm.com/support/pages/node/6998343
The list of APARs ("Authorized Program Analysis Reports")
does not contain anything that would affect security.

With this update, the Java-based
GUI requires a newer jdk version
(note: version 8.1.22.0 was not compatible with the newer jdk).

Also, all URLs in package and module comments are updated.